backend/requirements.txt
fastapi
pydantic
python-dotenv
transformers[torch]
langid